Refactor: move findType into FlatbuffersTypeReference.kt
FlatbuffersPsiImplUtil.kt is used by the grammar to be
visible to grammarkit for things like mixins. Since these
functions are only used as static functions for one class,
make them private static functions of that class
